adicionado put de endereço e telefone

+++ src/main/java/com/vssfullstack/usuario/business/UsuarioService.java
@@ -2,10 +2,16 @@ package com.vssfullstack.usuario.business;
 
 
 import com.vssfullstack.usuario.business.converter.UsuarioConverter;
+import com.vssfullstack.usuario.business.dto.EnderecoDTO;
+import com.vssfullstack.usuario.business.dto.TelefoneDTO;
 import com.vssfullstack.usuario.business.dto.UsuarioDTO;
+import com.vssfullstack.usuario.infrastructure.entity.Endereco;
+import com.vssfullstack.usuario.infrastructure.entity.Telefone;
 import com.vssfullstack.usuario.infrastructure.entity.Usuario;
 import com.vssfullstack.usuario.infrastructure.exceptions.ConflictException;
 import com.vssfullstack.usuario.infrastructure.exceptions.ResourceNotFoundException;
+import com.vssfullstack.usuario.infrastructure.repository.EnderecoRepository;
+import com.vssfullstack.usuario.infrastructure.repository.TelefoneRepository;
 import com.vssfullstack.usuario.infrastructure.repository.UsuarioRepository;
 import com.vssfullstack.usuario.infrastructure.security.JwtUtil;
 import lombok.RequiredArgsConstructor;
@@ -17,6 +23,8 @@ import org.springframework.stereotype.Service;
 public class UsuarioService {
 
     private final UsuarioRepository usuarioRepository;
+    private final EnderecoRepository enderecoRepository;
+    private final TelefoneRepository telefoneRepository;
     private final UsuarioConverter usuarioConverter;
     private final PasswordEncoder passwordEncoder;
     private final JwtUtil jwtUtil;
@@ -86,11 +94,21 @@ public class UsuarioService {
 
     // Deleta usuario por email
     public void deletaUsuarioPorEmail(String email) {
+
+        if (!usuarioRepository.existsByEmail(email)) {
+            throw new ResourceNotFoundException("Email não encontrado.");
+        }
+
         usuarioRepository.deleteByEmail(email);
+
     }
 
     // Atualiza dados do usuario
     public UsuarioDTO atualizaDadosUsuario(String token, UsuarioDTO usuarioDTO) {
+
+        if (usuarioDTO == null) {
+            throw new ConflictException("Dados do usuário não podem ser nulos.");
+        }
 
         // Aqui busca o email do usuário através do token (email)
         String email = jwtUtil.extractUsername(token.substring(7));
@@ -108,6 +126,29 @@ public class UsuarioService {
 
         // Salva os dados do usuário convertido e pega os dados convertidos para usuarioDTO
         return usuarioConverter.paraUsuarioDTO(usuarioRepository.save(usuario));
+    }
+
+    // Atualiza endereço
+    public EnderecoDTO atualizaDadosEndereco(Long idEndereco, EnderecoDTO enderecoDTO) {
+
+        Endereco enderecoEntity = enderecoRepository.findById(idEndereco).orElseThrow(
+                () -> new ResourceNotFoundException("Id não encontrado" +idEndereco));
+
+        Endereco endereco = usuarioConverter.atualizaEndereco(enderecoDTO, enderecoEntity);
+
+        return  usuarioConverter.paraEnderecoDTO(enderecoRepository.save(endereco));
+
+    }
+
+    // Atualiza telefone
+    public TelefoneDTO atualizaDadosTelefone(Long idTelefone, TelefoneDTO telefoneDTO) {
+
+        Telefone telefoneEntity = telefoneRepository.findById(idTelefone).orElseThrow(
+                () -> new ResourceNotFoundException("Id não encontrado" + idTelefone));
+
+        Telefone telefone = usuarioConverter.atualizaTelefone(telefoneDTO, telefoneEntity);
+
+        return usuarioConverter.paraTelefoneDTO(telefoneRepository.save(telefone));
     }
